fixing broken test
This test was broken with the addition of the final rule where cells
could become live again. Previously the middle cell would stay dead, but
now becomes live due to the 3 live cells on the top row. Since this test
is specifically concerned with the case of a live cell staying alive,
I've changed the world to be a 1-D array to avoid the newly live cell.